pyproject.toml
[project]
name = "5-types-of-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.21",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.8",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "transformers>=4.56.1",
    "unstructured[pdf]>=0.18.14",
]
